infrastructure/cdk/requirements.txt
# AWS CDK dependencies for orb-integration-hub infrastructure
aws-cdk-lib==2.172.0
constructs>=10.0.0,<11.0.0

# Utilities
pyyaml>=6.0.0
